server/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
playwright==1.38.0
pluggy==1.3.0
pyee==9.0.4
pytest==7.4.2
pytest-base-url==2.0.0
pytest-playwright==0.4.2
python-slugify==8.0.1
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
